Fix issues with COS instance flow

+++ cloud/scope/powervs_cluster.go
@@ -1288,7 +1288,7 @@ func (s *PowerVSClusterScope) ReconcileCOSInstance() error {
 		return err
 	}
 	if cosServiceInstanceStatus != nil {
-		s.SetStatus(TransitGateway, infrav1beta2.ResourceReference{ID: cosServiceInstanceStatus.GUID, ControllerCreated: pointer.Bool(false)})
+		s.SetStatus(COSInstance, infrav1beta2.ResourceReference{ID: cosServiceInstanceStatus.GUID, ControllerCreated: pointer.Bool(false)})
 	} else {
 		// create COS service instance
 		cosServiceInstanceStatus, err = s.createCOSServiceInstance()
@@ -1296,7 +1296,7 @@ func (s *PowerVSClusterScope) ReconcileCOSInstance() error {
 			s.Error(err, "error creating cos service instance")
 			return err
 		}
-		s.SetStatus(TransitGateway, infrav1beta2.ResourceReference{ID: cosServiceInstanceStatus.GUID, ControllerCreated: pointer.Bool(true)})
+		s.SetStatus(COSInstance, infrav1beta2.ResourceReference{ID: cosServiceInstanceStatus.GUID, ControllerCreated: pointer.Bool(true)})
 	}
 
 	apiKey := os.Getenv("IBMCLOUD_API_KEY")
@@ -1313,7 +1313,9 @@ func (s *PowerVSClusterScope) ReconcileCOSInstance() error {
 	s.COSClient = cosClient
 
 	// check bucket exist in service instance
-	if exist, err := s.checkCOSBucket(); exist || err != nil {
+	if exist, err := s.checkCOSBucket(); exist {
+		return nil
+	} else if err != nil {
 		s.Error(err, "error checking cos bucket")
 		return err
 	}
@@ -1361,6 +1363,8 @@ func (s *PowerVSClusterScope) createCOSBucket() error {
 	// If bucket already exists, all good.
 	case s3.ErrCodeBucketAlreadyOwnedByYou:
 		return nil
+	case s3.ErrCodeBucketAlreadyExists:
+		return nil
 	default:
 		return fmt.Errorf("error creating COS bucket %w", err)
 	}
@@ -1395,12 +1399,10 @@ func (s *PowerVSClusterScope) createCOSServiceInstance() (*resourcecontrollerv2.
 	//	return nil, fmt.Errorf("error retrieving id info for powervs service %w", err)
 	//}
 
-	//TODO(karthik-k-n): add funciton to fetch name
 	target := "Global"
-	serviceInstanceName := fmt.Sprintf("%s-%s", s.InfraCluster().GetName(), "cosInstance")
 	// create service instance
 	serviceInstance, _, err := s.ResourceClient.CreateResourceInstance(&resourcecontrollerv2.CreateResourceInstanceOptions{
-		Name:           &serviceInstanceName,
+		Name:           s.GetServiceName(COSInstance),
 		Target:         &target,
 		ResourceGroup:  &resourceGroupID,
 		ResourcePlanID: pointer.String(cosResourcePlanID),
@@ -1748,6 +1750,10 @@ func (s *PowerVSClusterScope) DeleteCosInstance() error {
 				return nil
 			}
 			return fmt.Errorf("error fetching COS instance: %w", err)
+		}
+
+		if cosInstance != nil && (*cosInstance.State == "pending_reclamation" || *cosInstance.State == string(infrav1beta2.ServiceInstanceStateRemoved)) {
+			return nil
 		}
 
 		_, err = s.ResourceClient.DeleteResourceInstance(&resourcecontrollerv2.DeleteResourceInstanceOptions{
